requirements.txt

# conda deactivate
# conda remove --name env --all
# conda create -n env python=3.8
# conda activate env
# CUDA 11.8
# pip install torch==2.0.0 torchvision==0.15.1 torchaudio==2.0.1 --index-url https://download.pytorch.org/whl/cu118
# conda install -c iopath iopath
# pip install "git+https://github.com/facebookresearch/pytorch3d.git"
# pip install -r requirements.txt



##### pip install torch==1.13.0+cu116 torchvision==0.14.0+cu116 torchaudio==0.13.0 --extra-index-url https://download.pytorch.org/whl/cu116
##### pip3 install torch torchvision torchaudio --index-url https://download.pytorch.org/whl/cu118



# --------- pytorch --------- #
#  torch==1.13.0
#  torchvision==0.14.0
lightning>=2.0.0
torchmetrics>=0.11.4

# --------- hydra --------- #
hydra-core==1.2.0
hydra-colorlog==1.2.0
hydra-optuna-sweeper==1.2.0

# --------- loggers --------- #y

wandb~=0.13.5
# neptune-client
# mlflow
# comet-ml
# aim>=3.16.2  # no lower than 3.16.2, see https://github.com/aimhubio/aim/issues/2550

# --------- others --------- #
rootutils       # standardizing the project root setup
pre-commit      # hooks for applying linters on commit
rich            # beautiful text formatting in terminal
pytest          # tests
# sh            # for running bash commands in some tests (linux/macos only)

omegaconf~=2.2.3
numpy~=1.23.4
joblib~=1.1.1
matplotlib~=3.5.3
scikit-learn~=1.1.3
lmdb~=1.3.0
scipy~=1.8.1
librosa~=0.9.2
tqdm~=4.64.1
pandas~=1.5.1
mysql-connector

packaging~=21.3
einops~=0.6.0
ipython~=8.6.0

google-cloud-texttospeech
pygame
torch-audiomentations
transformers
seaborn

# conda install -c conda-forge ffmpeg
